// src/Umbraco.Cms.Search.Core.Client/Client/src/settings/api/types.gen.ts
export type HealthStatusModel = 'Healthy' | 'Rebuilding' | 'Corrupted' | 'Empty' | 'Unknown';

export type IndexModel = {
    indexAlias: string;
    providerName: string;
    documentCount: number;
    healthStatus: HealthStatusModel;
};